Split model.go into a file per type
Every other type in rest/model already lives in a file named after it, which
is how you find one. model.go held thirteen unrelated types, so anything from
a backup request to the generic response envelope was somewhere in one
alphabetically meaningless file.

Each type now has its own file, named after it, like the ones already there.
Access keeps its three predicates alongside it. No type, field or tag changed.
