+++ backend/rest/model/activation_status_response.go
@@ -1,0 +1,6 @@
+package model
+
+type ActivationStatusResponse struct {
+	Activated bool   `json:"activated"`
+	DeviceUrl string `json:"device_url"`
+}
