+++ backend/rest/model/backup_create_request.go
@@ -1,0 +1,5 @@
+package model
+
+type BackupCreateRequest struct {
+	App string `json:"app"`
+}
