+++ backend/rest/model/backup_remove_request.go
@@ -1,0 +1,5 @@
+package model
+
+type BackupRemoveRequest struct {
+	File string `json:"file"`
+}
