+++ backend/rest/model/backup_restore_request.go
@@ -1,0 +1,5 @@
+package model
+
+type BackupRestoreRequest struct {
+	File string `json:"file"`
+}
