+++ backend/rest/model/event_trigger_request.go
@@ -1,0 +1,5 @@
+package model
+
+type EventTriggerRequest struct {
+	Event string `json:"event"`
+}
