+++ backend/rest/model/mail_relay.go
@@ -1,0 +1,5 @@
+package model
+
+type MailRelay struct {
+	Enabled bool `json:"enabled"`
+}
