+++ backend/rest/model/mail_relay_credentials.go
@@ -1,0 +1,9 @@
+package model
+
+type MailRelayCredentials struct {
+	Enabled  bool   `json:"enabled"`
+	Host     string `json:"host,omitempty"`
+	Port     int    `json:"port,omitempty"`
+	Login    string `json:"login,omitempty"`
+	Password string `json:"password,omitempty"`
+}
