+++ backend/rest/model/redirect_info_response.go
@@ -1,0 +1,5 @@
+package model
+
+type RedirectInfoResponse struct {
+	Domain string `json:"domain"`
+}
