+++ backend/rest/model/storage_activate_disks_request.go
@@ -1,0 +1,6 @@
+package model
+
+type StorageActivateDisksRequest struct {
+	Devices []string `json:"devices"`
+	Format  bool     `json:"format"`
+}
