+++ backend/rest/model/storage_activate_partition_request.go
@@ -1,0 +1,6 @@
+package model
+
+type StorageActivatePartitionRequest struct {
+	Device string `json:"device"`
+	Format bool   `json:"format"`
+}
