+++ backend/rest/model/user_info.go
@@ -1,0 +1,6 @@
+package model
+
+type UserInfo struct {
+	Admin    bool   `json:"admin"`
+	Username string `json:"username"`
+}
